Bert

+++ Server_Banking/sql/226team_saving.sql
@@ -25,7 +25,7 @@ DROP TABLE IF EXISTS `saving`;
 CREATE TABLE `saving` (
   `account_num` varchar(255) NOT NULL,
   `customer` int(11) NOT NULL,
-  `balance` int(11) NOT NULL,
+  `balance` decimal(11,0) NOT NULL,
   PRIMARY KEY (`account_num`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 /*!40101 SET character_set_client = @saved_cs_client */;
@@ -49,4 +49,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2019-11-26  1:37:45
+-- Dump completed on 2019-11-26 14:59:13
